requirement.txt
Flask-2.3.2
Jinja2-3.1.2
MarkupSafe-2.1.3
Werkzeug-2.3.6
blinker-1.6.2
click-8.1.3
colorama-0.4.6
itsdangerous-2.1.2
pymysql-1.0.3